requirements.txt
numpy
pandas
ffmpeg-python
streamlit
requests
flask
nodejs-bin[cmd]